Functions

PotentialLearning.GaN — Type

GaN Potential
See https://iopscience.iop.org/article/10.1088/1361-648X/ab6cbe

PotentialLearning.GaN — Method

GaN(params::Dict)

Creates

PotentialLearning.LennardJones — Type

Lennard-Jones Potential

PotentialLearning.SNAP_LAMMPS — Type

Wrapper of the SNAP implementation of LAMMPS, built with LAMMPS.jl
Mathematical formulation: A. P. Thompson et al.
                          http://dx.doi.org/10.1016/j.jcp.2014.12.018

PotentialLearning.SNAP_LAMMPS — Method

SNAP_LAMMPS(params::Dict)

Creation of a SNAP_LAMMPS instance based on the configuration parameters

PotentialLearning.calc_A — Method

calc_A(path::String, p::SNAP_LAMMPS)

Calculation of the A matrix of SNAP (Eq. 13, 10.1016/j.jcp.2014.12.018)

PotentialLearning.error — Method

error(β::Vector{Float64}, p, s::SNAP_LAMMPS)

Error function to perform the learning process (Eq. 14, 10.1016/j.jcp.2014.12.018)

PotentialLearning.learn — Method

learn(p::Potential, dft_training_data::Vector{Float64}, params::Dict)

Fit the potentials, forces, and stresses against the DFT data using the configuration parameters.

PotentialLearning.load_conf_params — Method

load_conf_params(path::String)

Load configuration parameters

PotentialLearning.load_dft_data — Method

load_dft_data(params::Dict)

Load DFT data

PotentialLearning.load_positions_per_conf — Method

load_positions_per_conf(path::String, no_atoms_per_conf::Int64,
                             no_conf_init::Int64, no_conf_end::Int64)

Load atomic positions per configuration

PotentialLearning.potential_energy — Method

potential_energy(params::Dict, j::Int64, p::Potential)

Calculation of the potential energy of a particular atomic configuration (j). This calculation requires accessing the SNAP implementation of LAMMPS.

PotentialLearning.potential_energy — Method

potential_energy(atomic_positions::Vector{Position}, rcut::Float64, p::Potential)

Calculation of the potential energy of a particular atomic configuration. It is based on the atomic positions of the configuration, the rcut, and a particular potential.

PotentialLearning.validate — Method

validate(p::Potential, dft_validation_data::Vector{Float64}, params::Dict)

Validate trained potentials, forces, and stresses.