Functions

PotentialLearning.BornMayer — Type

Born-Mayer Potential

PotentialLearning.BornMayer — Method

BornMayer(params::Dict)

Creates a BM potential.

PotentialLearning.Coulomb — Type

Coulomb Potential

PotentialLearning.Coulomb — Method

Coulomb(params::Dict)

Creates a Coulomb potential.

PotentialLearning.GaN — Type

GaN Potential

See 10.1088/1361-648X/ab6cbe

PotentialLearning.LennardJones — Type

Lennard-Jones Potential

PotentialLearning.LennardJones — Method

LennardJones(params::Dict)

Creates the LennardJones potential.

PotentialLearning.SNAP_LAMMPS — Type

SNAP_LAMMPS is a PotentialLearningProblem, it is based on the SNAP 
implementation of LAMMPS, which is accessed through LAMMPS.jl.
Mathematical formulation: A. P. Thompson et al. (10.1016/j.jcp.2014.12.018)

PotentialLearning.SNAP_LAMMPS — Method

SNAP_LAMMPS(dft_data::Vector{Float64}, ref_data::Vector{Float64}, params::Dict)

Creates a SNAP_LAMMPS.

PotentialLearning.ZBL — Type

Ziegler-Biersack-Littmark (ZBL) Potential

See https://docs.lammps.org/pair_zbl.html

PotentialLearning.ZBL — Method

ZBL(params::Dict)

Creates a ZBL potential.

PotentialLearning.S — Method

S(r)

Switching function that ramps the energy, force, and curvature smoothly to zero between an inner and outer cutoff. Here, the inner and outer cutoff are the same for all pairs of atom types.

PotentialLearning.atom_type — Method

atom_type(i::Int64)

Returns the atom type of the i-th atom of the configuration.

PotentialLearning.calc_A — Method

calc_A(params::Dict, p::SNAP_LAMMPS)

Calculates the matrix A (See Eq. 13, 10.1016/j.jcp.2014.12.018). This calculation requires accessing the SNAP implementation of LAMMPS.

PotentialLearning.error — Method

error(β::Vector{Float64}, s::SNAP_LAMMPS)

Error function needed to perform the learning process.

PotentialLearning.error_metrics — Method

error_metrics(p::PotentialLearningProblem, val_data::Vector{Float64}, params::Dict)

Calculates error metrics: maxrelerror, mae, rmse

PotentialLearning.forces — Method

forces(p::Potential, atomic_positions::Vector{Position}, rcutfac::Float64)

Calculates the forces of each atom in an atomic configuration.

PotentialLearning.forces — Method

forces(p::Potential, j::Int64)

Calculates the forces of a particular atomic configuration (j) using the fitted parameters β. This calculation requires accessing the SNAP implementation of LAMMPS. See https://docs.lammps.org/computesnaatom.html#compute-snad-atom-command

PotentialLearning.generate_data — Method

generate_data(params::Dict)

Generates DFT/reference data based on a particular potential.

PotentialLearning.generate_data_aux — Method

generate_data_aux(p::Potential, positions_per_conf::Vector,
                  a::Int64, b::Int64, rcutfac::Float64, fit_forces::Bool)

Auxiliar function. See generate_data.

PotentialLearning.get_bispectrums — Method

get_bispectrums(lmp, data_path::String, p::SNAP_LAMMPS)

Calculates the bispectrums components and its derivatives. See https://docs.lammps.org/computesnaatom.html

PotentialLearning.get_conf_params — Method

get_conf_params(path::String)

Loads configuration parameters.

PotentialLearning.learn — Method

learn(p::PotentialLearningProblem, params::Dict)

Fits the potentials, forces, and stresses against the DFT and reference data using the configuration parameters.

PotentialLearning.learning_problem — Method

learning_problem( dft_train_data::Vector{Float64},
                  ref_train_data::Vector{Float64},
                  params::Dict)

Creates a potential learning problem.

PotentialLearning.linearize — Method

linearize(data::Vector{Force})

Linearizes a vector of forces. E.g. [Force(1.0,2.0,3.0), Force(4.0,5.0,6.0)] => [1.0, 2.0, 3.0, 4.0, 5.0, 6.0]

PotentialLearning.load_positions_per_conf — Method

load_positions_per_conf(path::String, no_atoms_per_conf::Int64,
                        no_conf_init::Int64, no_conf_end::Int64)

Loads atomic positions per configuration.

PotentialLearning.potential_energy — Method

potential_energy(p::BornMayer, r::Position, args...)

Calculates Born-Mayer potential energy.

PotentialLearning.potential_energy — Method

potential_energy(p::Coulomb, r::Position, args...)

Calculates Coulomb potential energy.

PotentialLearning.potential_energy — Method

potential_energy(p::GaN, r::Position, args...)

Calculates GaN potential energy.

PotentialLearning.potential_energy — Method

potential_energy(p::LennardJones, r::Position, args...)

Calculates LennardJones potential energy.

PotentialLearning.potential_energy — Method

potential_energy(p::Potential, atomic_positions::Vector{Position}, rcutfac::Float64)

Calculates the potential energy of a particular atomic configuration. It is based on the atomic positions of the configuration, the rcutfac, and a particular potential.

PotentialLearning.potential_energy — Method

potential_energy(p::Potential, j::Int64)

Calculates the potential energy of a particular atomic configuration (j) using the fitted parameters β. This calculation requires accessing the SNAP implementation of LAMMPS.

PotentialLearning.potential_energy — Method

potential_energy(p::ZBL, r::Position, args...)

Calculates the potential energy of ZBL with a switching function.

PotentialLearning.zbl — Method

zbl(p::ZBL, r::Float64)

ZBL function without the switching function S.

PotentialLearning.ϕ — Method

ϕ(p::ZBL, x::Float64)

Auxiliary function, necessary to calculate ZBL.