API Reference

This page provides a list of all documented types and functions and in PotentialLearning.jl.

PotentialLearning.AbstractLearningProblem — Type

LearningProblem{T<:Real} <: AbstractLearningProblem 
    ds      :: DataSet
    logprob :: Function # logprob(y::params, ds::Dataset) -> scalar
    ∇logprob :: Function # ∇logprob(y::params, ds::Dataset) -> gradient wrt y (params)
    params :: Vector{T} # model parameters

Generic LearningProblem that allows the user to pass a logprob(y::params, ds::DataSet) function and its gradient. The gradient should return a vector of logprob with respect to it's params. If the user does not have a gradient function available, then Flux can provide one for it (provided that logprob is of the form above).

PotentialLearning.ActiveSubspace — Type

ActiveSubspace{T<:Real} <: DimensionReducer
    Q :: Function 
    ∇Q :: Function (gradient of Q)
    tol :: T

Use the theory of active subspaces, with a given quantity of interest (expressed as the function Q) which takes a Configuration as an input and outputs a real scalar. ∇Q should input a Configuration and output an appropriate gradient. If tol is a float then the number of components to keep is determined by the smallest n such that relative percentage of variance explained by keeping the leading n principle components is greater than 1 - tol. If tol is an int, then we return the components corresponding to the tol largest eigenvalues.

PotentialLearning.AtomicData — Type

AtomicData <: Data

Abstract type declaring the type of information that is unique to a particular atom (instead of a whole configuration).

PotentialLearning.Configuration — Method

Configuration(data::Union{AtomsBase.FlexibleSystem, ConfigurationData} )

A Configuration is a data struct that contains information unique to a particular configuration of atoms (Energy, LocalDescriptors, ForceDescriptors, and a FlexibleSystem) in a dictionary. Example: '''julia e = Energy(-0.57, u"eV") ld = LocalDescriptors(...) c = Configuration(e, ld) '''

Configurations can be added together, which merges the data dictionaries '''julia c1 = Configuration(e) # Contains energy c2 = Configuration(f) # contains forces c = c1 + c2 # c <: Configuration, contains energy and forces '''

PotentialLearning.ConfigurationData — Type

ConfigurationData <: Data

Abstract type declaring the type of data that is unique to a particular configuration (instead of just an atom).

PotentialLearning.CorrelationMatrix — Type

CorrelationMatrix 
    α :: Vector{Float64} # weights

CorrelationMatrix produces a global descriptor that is the correlation matrix of the local descriptors. In other words, it is mean(bi'*bi for bi in B).

PotentialLearning.CovariateLinearProblem — Type

CovariateLinearProblem{T<:Real} <: LinearProblem{T}
    e       :: Vector
    f       :: Vector{Vector{T}}
    B       :: Vector{Vector{T}}
    dB      :: Vector{Matrix{T}}
    β       :: Vector{T} 
    σe      :: Vector{T} 
    σf      :: Vector{T}
    Σ       :: Symmetric{T, Matrix{T}}

A CovariateLinearProblem is a linear problem in which we are fitting energies and forces using both descriptors and their gradients (B and dB, respectively). When this is the case, the solution is not available analytically and must be solved using some iterative optimization proceedure. In the end, we fit the model coefficients, β, standard deviations corresponding to energies and forces, σe and σf, and the covariance Σ.

PotentialLearning.Data — Type

Data

Abstract supertype of ConfigurationData.

PotentialLearning.DataBase — Type

DataBase

Abstract type for DataSets.

PotentialLearning.DataSet — Type

DataSet

Struct that holds vector of configuration. Most operations in PotentialLearning are built around the DataSet structure.

PotentialLearning.Distance — Type

Distance

A struct of abstract type Distance produces the distance between two `global` descriptors, or features. Not all distances might be compatible with all types of features.

PotentialLearning.DotProduct — Type

DotProduct <: Kernel 
    α :: Power of DotProduct kernel 

Computes the dot product kernel between two features, i.e.,

cos(θ) = ( A ⋅ B / (||A||^2||B||^2) )^α

PotentialLearning.Energy — Type

Energy <: ConfigurationData
    d :: Real
    u :: Unitful.FreeUnits

Convenience struct that holds energy information (and corresponding units). Default unit is eV

PotentialLearning.Euclidean — Type

Euclidean <: Distance 
    Cinv :: Covariance Matrix 

Computes the squared euclidean distance with weight matrix Cinv, the inverse of some covariance matrix.

PotentialLearning.ExtXYZ — Type

ExtXYZ <: IO

PotentialLearning.Feature — Type

Feature

A struct of abstract type Feature represents a function that takes in a set of local descriptors corresponding to some atomic environment and produce a global descriptor.

PotentialLearning.Force — Type

Force <: AtomicData 
    f :: Vector{<:Real}
    u :: Unitful.FreeUnits

Contains the force with (x,y,z)-components in f with units u. Default unit is "eV/Å".

PotentialLearning.ForceDescriptor — Type

ForceDescriptor <: AtomicData
    b :: Vector{<:Vector{<:Real}}

Contains the x,y,z components (out vector) of the force descriptor (inner vector).

PotentialLearning.ForceDescriptors — Type

ForceDescriptors <: ConfigurationData
    b :: Vector{ForceDescriptor}

A container holding all of the ForceDescriptors for all atoms in a configuration.

PotentialLearning.Forces — Type

Forces <: ConfigurationData
    f :: Vector{force}

Forces is a struct that contains all force information in a configuration.

PotentialLearning.Forstner — Type

Forstner <: Distance 
    α :: Regularization parameter

Computes the squared Forstner distance between two positive semi-definite matrices.

PotentialLearning.GlobalMean — Type

    GlobalMean{T}

GlobalMean produces the mean of the local descriptors.

PotentialLearning.GlobalSum — Type

    GlobalSum{T}

GlobalSum produces the sum of the local descriptors.

PotentialLearning.Kernel — Type

Kernel

A struct of abstract type Kernel is function that takes in two features and produces a semi-definite scalar representing the similarity between the two features.

PotentialLearning.LAMMPS — Type

struct LAMMPS <: IO
    elements :: Vector{Symbol}
    boundary_conditions :: Vector
end

PotentialLearning.LinearProblem — Type

LinearPRoblem{T<:Real} <: AbstractLearningProblem

An abstract type to specify linear potential inference problems.

PotentialLearning.LinearProblem — Method

LinearProblem(ds::DatasSet; T = Float64)

Construct a LinearProblem by detecting if there are energy descriptors and/or force descriptors and construct the appropriate LinearProblem (either Univariate, if only a single type of descriptor, or Covariate, if there are both types).

PotentialLearning.LocalDescriptor — Type

LocalDescriptor <: AtomicData

A vector corresponding to the descriptor for a particular atom's neighborhood.

PotentialLearning.LocalDescriptors — Type

LocalDescriptors <: ConfigurationData

A vector of LocalDescriptor, which now should represent all local descriptors for atoms in a configuration.

PotentialLearning.PCA — Type

PCA <: DimensionReducer
    tol :: Float64

Use SVD to compute the PCA of the design matrix of descriptors. (using Force descriptors TBA)

If tol is a float then the number of components to keep is determined by the smallest n such that relative percentage of variance explained by keeping the leading n principle components is greater than 1 - tol. If tol is an int, then we return the components corresponding to the tol largest eigenvalues.

PotentialLearning.RBF — Type

RBF <: Kernel 
    d :: Distance function 
    α :: Reguarlization parameter 
    ℓ :: Length-scale parameter
    β :: Scale parameter

Computes the squared exponential kernel, i.e.,

 k(A, B) = β xp( -rac{1}{2} d(A,B)/ℓ^2 ) + α δ(A, B)

PotentialLearning.RandomSelector — Type

struct Random
    num_configs :: Int 
    batch_size  :: Int 
end

A convenience function that allows the user to randomly select indices uniformly over [1, num_configs].

PotentialLearning.UnivariateLinearProblem — Type

UnivariateLinearProblem{T<:Real} <: LinearProblem{T}
    iv_data :: Vector 
    dv_data :: Vector 
    β       :: Vector{T}
    σ       :: Vector{T} 
    Σ       :: Symmetric{T, Matrix{T}}
end

A UnivariateLinearProblem is a linear problem in which there is only 1 type of independent variable / dependent variable. Typically, that means we are either only fitting energies or only fitting forces. When this is the case, the solution is available analytically and the standard deviation, σ, and covariance, Σ, of the coefficients, β, are computable.

PotentialLearning.YAML — Type

YAML <: IO
    energy_units :: Unitful.FreeUnits
    distance_units :: Unitful.FreeUnits

PotentialLearning.kDPP — Type

struct kDPP
    K :: EllEnsemble
end

A convenience function that allows the user access to a k-Determinantal Point Process through Determinantal.jl. All that is required to construct a kDPP is a similarity kernel, for which the user must provide a LinearProblem and two functions to compute descriptor (1) diversity and (2) quality.

PotentialLearning.kDPP — Method

kDPP(ds::Dataset, f::Feature, k::Kernel)

A convenience function that allows the user access to a k-Determinantal Point Process through Determinantal.jl. All that is required to construct a kDPP is a dataset, a method to compute features, and a kernel. Optional arguments include batch size and type of descriptor (default LocalDescriptors).

PotentialLearning.kDPP — Method

kDPP(features::Union{Vector{Vector{T}}, Vector{Symmetric{T, Matrix{T}}}}, k::Kernel)

A convenience function that allows the user access to a k-Determinantal Point Process through Determinantaljl. All that is required to construct a kDPP are features (either a vector of vector features or a vector of symmetric matrix features) and a kernel. Optional argument is batch_size (default length(features)).

PotentialLearning.KernelMatrix — Method

KernelMatrix(ds1::DataSet, ds2::DataSet, F::Feature, k::Kernel)

Compute nonsymmetric kernel matrix K using features of the datasets ds1 and ds2 calculated using the Feature method F.

PotentialLearning.KernelMatrix — Method

KernelMatrix(ds::DataSet, F::Feature, k::Kernel)

Compute symmetric kernel matrix K using features of the dataset ds calculated using the Feature method F.

PotentialLearning.KernelMatrix — Method

KernelMatrix(F, k::Kernel)

Compute symmetric kernel matrix K where K{ij} = k(Fi, F_j).

PotentialLearning.KernelMatrix — Method

KernelMatrix(F1, F2, k::Kernel)

Compute non-symmetric kernel matrix K where K{ij} = k(F1i, F2_j).

PotentialLearning.calc_metrics — Method

calc_metrics(x_pred, x)

x_pred: vector of predicted values of a variable. E.g. energy. x: vector of true values of a variable. E.g. energy.

Returns MAE, RMSE, and RSQ.

PotentialLearning.compute_features — Method

compute_feature(ds::DataSet, f::Feature; dt = LocalDescriptors)

Computes features of the dataset ds using the feature method F on descriptors dt (default option are the LocalDescriptors, if available).

PotentialLearning.compute_kernel — Method

compute_kernel(A, B, k)

Compute similarity kernel between features A and B using kernel k.

PotentialLearning.fit — Function

fit(ds::DataSet, dr::DimensionReducer)

Fits a linear dimension reduction routine using information from DataSet. See individual types of DimensionReducers for specific details.

PotentialLearning.fit — Method

fit(ds::DataSet, as::ActiveSubspace)

Fits a linear dimension reduction routine using the eigendirections of the uncentered covariance of the function ∇Q(c::Configuration) over the configurations in ds. Primarily used to reduce the dimension of the descriptors.

PotentialLearning.fit — Method

fit(ds::DataSet, pca::PCA)

Fits a linear dimension reduction routine using PCA on the global descriptors in the dataset ds.

PotentialLearning.fit_transform — Method

fit_transform(ds::DataSet, dr::DimensionReducer)

Fits a linear dimension reduction routine using information from DataSet and performs dimension reduction on descriptors and force_descriptors (whichever are available). See individual types of DimensionReducers for specific details.

PotentialLearning.get_batches — Method

get_batches(n_batches, B_train, B_train_ext, e_train, dB_train, f_train,
            B_test, B_test_ext, e_test, dB_test, f_test)

n_batches: no. of batches per dataset. B_train: descriptors of the energies used in training. B_train_ext: extendended descriptors of the energies used in training. Requiered to compute forces. e_train: energies used in training. dB_train: derivatives of the energy descritors used in training. f_train: forces used in training. B_test: descriptors of the energies used in test. B_test_ext: extendended descriptors of the energies used in test. Requiered to compute forces. e_test: energies used in test. dB_test: derivatives of the energy descritors used in test. f_test: forces used in test.

Returns the data loaders for training and test of energies and forces.

PotentialLearning.get_dpp_mode — Method

get_dpp_mode(dpp::kDPP, batch_size::Int) <: Vector{Int64}

Access an approximate mode of the k-DPP as calculated by a greedy subset algorithm. See Determinantal.jl for details.

PotentialLearning.get_energy — Method

get_energy(c::Configuration) <: Energy

Retrieves the energy (if available) in the Configuration c.

PotentialLearning.get_force_descriptors — Method

get_force_descriptors(c::Configuration) <: ForceDescriptors

Retrieves the force descriptors (if available) in the Configuration c.

PotentialLearning.get_forces — Method

get_forces(c::Configuration) <: Forces

Retrieves the forces (if available) in the Configuration c.

PotentialLearning.get_inclusion_prob — Method

get_inclusion_prob(dpp::kDPP) <: Vector{Float64}

Access an approximation to the inclusion probabilities as calculated by Determinantal.jl (see package for details).

PotentialLearning.get_input — Method

get_input(args)

args: vector of arguments (strings)

Returns an OrderedDict with the arguments. See https://github.com/cesmix-mit/AtomisticComposableWorkflows documentation for information about how to define the input arguments.

PotentialLearning.get_local_descriptors — Method

get_local_descriptors(c::Configuration) <: LocalDescriptors

Retrieves the local descriptors (if available) in the Configuration c.

PotentialLearning.get_metrics — Method

get_metrics( e_train_pred, e_train, f_train_pred, f_train,
             e_test_pred, e_test, f_test_pred, f_test,
             B_time, dB_time, time_fitting)

e_train_pred: vector of predicted training energy values. e_train: vector of true training energy values. f_train_pred: vector of predicted training force values. f_train: vector of true training force values. e_test_pred: vector of predicted test energy values. e_test: vector of true test energy values. f_test_pred: vector of predicted test force values. f_test: vector of true test force values. B_time: elapsed time consumed by descriptors calculation. dB_time: elapsed time consumed by descriptor derivatives calculation. time_fitting: elapsed time consumed by fitting process.

Computes MAE, RMSE, and RSQ for training and testing energies and forces. Also add elapsed times about descriptors and fitting calculations. Returns an OrderedDict with the information above.

PotentialLearning.get_metrics — Method

get_metrics( e_train_pred, e_train,
             e_test_pred, e_test)

e_train_pred: vector of predicted training energy values. e_train: vector of true training energy values. e_test_pred: vector of predicted test energy values. e_test: vector of true test energy values.

Computes MAE, RMSE, and RSQ for training and testing energies (no forces). Returns an OrderedDict with the information above.

PotentialLearning.get_positions — Method

get_positions(c::Configuration) <: Vector{SVector}

Retrieves the AtomsBase system positions (if available) in the Configuration c.

PotentialLearning.get_random_subset — Function

get_random_subset(r::Random, batch_size :: Int) <: Vector{Int64}

Access a random subset of the data as sampled from the provided k-DPP. Returns the indices of the random subset and the subset itself.

PotentialLearning.get_random_subset — Method

get_random_subset(dpp::kDPP, batch_size :: Int) <: Vector{Int64}

Access a random subset of the data as sampled from the provided k-DPP. Returns the indices of the random subset and the subset itself.

PotentialLearning.get_system — Method

get_system(c::Configuration) <: AtomsBase.AbstractSystem

Retrieves the AtomsBase system (if available) in the Configuration c.

PotentialLearning.get_values — Method

get_values(e::Energy) <: Real

Get the underlying real value (= e.d)

PotentialLearning.learn! — Method

learn!(lb::LBasisPotential, ds::DataSet, ss::SubsetSelector; α = 1e-8, return_cov = true)

Fits the linear basis potential lb using data from DataSet using batch gradient descent with batches provided by the subset selector ss. Optionally returns the posterior covariance over the parameters.

PotentialLearning.learn! — Method

learn!(lb::LBasisPotential, ds::DataSet; α = 1e-8, return_cov = true)

Fits the linear basis potential lb using data from DataSet using either an analytic solution or an optimization proceedure, depending on whether both energies and forces are present. Optionally returns the posterior covariance over the parameters.

PotentialLearning.learn! — Method

learn!(lp::CovariateLinearProblem, ss::SubsetSelector; α = 1e-8)

Fit a Gaussian distribution by finding the MLE of the following log probability: ℓ(β, σe, σf) = -0.5(e - A_e *β)'(e - Ae * β) / σe - 0.5*(f - Af β)'(f - A_f * β) / σf - log(σe) - log(σf)

through an iterative batch gradient descent optimization proceedure where the batches are provided by the subset selector.

PotentialLearning.learn! — Method

learn!(lp::CovariateLinearProblem; α = 1e-8)

Fit a Gaussian distribution by finding the MLE of the following log probability: ℓ(β, σe, σf) = -0.5(e - A_e *β)'(e - Ae * β) / σe - 0.5*(f - Af β)'(f - A_f * β) / σf - log(σe) - log(σf)

through an optimization procedure.

PotentialLearning.learn! — Method

learn!(lp::LearningProblem, ss::SubsetSelector; num_steps = 100 :: Int, opt = Flux.Optimisers.Adam())

Attempts to fit the parameters lp.params in the learning problem lp using batch gradient descent with the optimizer opt and num_steps number of iterations. Batching is provided by the passed ss::SubsetSelector.

PotentialLearning.learn! — Method

learn!(lp::LearningProblem;; num_steps = 100 :: Int, opt = Flux.Optimisers.Adam())

Attempts to fit the parameters lp.params in the learning problem lp using gradient descent with the optimizer opt and num_steps number of iterations.

PotentialLearning.learn! — Method

learn!(lp::UnivariateLinearProblem, ss::SubsetSelector; α = 1e-8)

Fit a univariate Gaussian distribution for the equation y = Aβ + ϵ, where β are model coefficients and ϵ ∼ N(0, σ). Fitting is done via batched gradient descent with batches provided by the subset selector and the gradients are calculated using Flux.

PotentialLearning.learn! — Method

learn!(lp::UnivariateLinearProblem; α = 1e-8)

Fit a univariate Gaussian distribution for the equation y = Aβ + ϵ, where β are model coefficients and ϵ ∼ N(0, σ). Fitting is done via SVD on the design matrix, A'*A (formed iteratively), where eigenvalues less than α are cut-off.

PotentialLearning.linearize_forces — Method

linearize_forces(forces)

forces: vector of forces per system

Returns a vector with the components of the forces of the systems.

PotentialLearning.load_data — Method

load_data(file::string, extxyz::ExtXYZ)
Load configuration from an extxyz file into a DataSet

PotentialLearning.load_data — Method

load_data(file::string, yaml::YAML)

Load configurations from a yaml file into a Vector of Flexible Systems, with Energies and Force.
Returns 
    ds - DataSet
    t = Vector{Dict} (any miscellaneous info from yaml file)

PotentialLearning.load_datasets — Method

load_datasets(input)

input: OrderedDict with input arguments. See get_defaults_args().

Returns training and test systems, energies, forces, and stresses.

PotentialLearning.to_num — Method

to_num(str)

str: string with a number: integer or float

Returns an integer or float.