API Reference

This page provides a list of all documented types and functions and in PotentialLearning.jl.

PotentialLearning.NNBasisPotential — Type

NNBasisPotential

Definition of the neural network basis potential composed type.

PotentialLearning.batch_train_flux! — Method

batch_train_flux!(train_loader_e, train_loader_f, w_e, w_f, ps, re, opt)

train_loader_e: energy data loader. train_loader_f: force data loader. w_e: energy weight. w_f: force weight. ps: neural network parameters. See Flux.destructure. re: neural network restructure. See Flux.destructure. opt: optimizer.

Batch training using Flux.jl. Returns neural network parameters.

PotentialLearning.batch_train_opt! — Method

batch_train_opt!(train_loader_e, train_loader_f, w_e, w_f, 
                 ps, re, epoch, opt, maxiters, train_losses_batches)

train_loader_e: energy data loader. train_loader_f: force data loader. w_e: energy weight. w_f: force weight. ps: neural network parameters. See Flux.destructure. re: neural network restructure. See Flux.destructure. epoch: current epoch. opt: optimizer. maxiters: maximum number of iterations in the optimizer. train_losses_batches: vector of batch losses during training.

Batch training using Optimization.jl. Returns NN parameters.

PotentialLearning.calc_metrics — Method

calc_metrics(x_pred, x)

x_pred: vector of predicted values of a variable. E.g. energy. x: vector of true values of a variable. E.g. energy.

Returns MAE, RMSE, and RSQ.

PotentialLearning.force — Method

force(A::AbstractSystem, p::NNBasisPotential)

A: atomic system or atomic configuration. p: neural network basis potential.

Returns the force of a system using a neural network basis potential.

PotentialLearning.force — Method

force(b::Vector, dbdr::Vector, p::NNBasisPotential)

b: energy descriptors of a system. dbdr: force descriptors of a system. p: neural network basis potential.

Returns the force of a system using a neural network basis potential.

PotentialLearning.force — Method

force(b::Vector, dbdr::Vector, ps::Vector, re)

b: energy descriptors of a system. dbdr: force descriptors of a system. ps: neural network parameters. See Flux.destructure. re: neural network restructure. See Flux.destructure.

Returns the force of a system using a neural network basis potential.

PotentialLearning.get_batches — Method

get_batches(n_batches, B_train, B_train_ext, e_train, dB_train, f_train,
            B_test, B_test_ext, e_test, dB_test, f_test)

n_batches: no. of batches per dataset. B_train: descriptors of the energies used in training. B_train_ext: extendended descriptors of the energies used in training. Requiered to compute forces. e_train: energies used in training. dB_train: derivatives of the energy descritors used in training. f_train: forces used in training. B_test: descriptors of the energies used in test. B_test_ext: extendended descriptors of the energies used in test. Requiered to compute forces. e_test: energies used in test. dB_test: derivatives of the energy descritors used in test. f_test: forces used in test.

Returns the data loaders for training and test of energies and forces.

PotentialLearning.get_input — Method

get_input(args)

args: vector of arguments (strings)

Returns an OrderedDict with the arguments. See https://github.com/cesmix-mit/AtomisticComposableWorkflows documentation for information about how to define the input arguments.

PotentialLearning.get_metrics — Method

get_metrics( e_train_pred, e_train, f_train_pred, f_train,
             e_test_pred, e_test, f_test_pred, f_test,
             B_time, dB_time, time_fitting)

e_train_pred: vector of predicted training energy values. e_train: vector of true training energy values. f_train_pred: vector of predicted training force values. f_train: vector of true training force values. e_test_pred: vector of predicted test energy values. e_test: vector of true test energy values. f_test_pred: vector of predicted test force values. f_test: vector of true test force values. B_time: elapsed time consumed by descriptors calculation. dB_time: elapsed time consumed by descriptor derivatives calculation. time_fitting: elapsed time consumed by fitting process.

Computes MAE, RMSE, and RSQ for training and testing energies and forces. Also add elapsed times about descriptors and fitting calculations. Returns an OrderedDict with the information above.

PotentialLearning.global_energy_loss — Method

global_energy_loss(loader_e, w_e, ps, re)

loader_e: preditected energies w_e: energy weight ps: neural network parameters. See Flux.destructure. re: neural network restructure. See Flux.destructure.

Returns the weighted global (all batches) loss of the energies.

PotentialLearning.global_force_loss — Method

global_force_loss(loader_f, w_f, ps, re)

loader_f: energies w_f: force weight ps: neural network parameters. See Flux.destructure. re: neural network restructure. See Flux.destructure.

Returns the weighted global (all batches) loss of the forces.

PotentialLearning.global_loss — Method

global_loss(loader_e, loader_f, w_e, w_f, ps, re)

loader_e: preditected energies loader_f: energies w_e: energy weight w_f: force weight ps: neural network parameters. See Flux.destructure. re: neural network restructure. See Flux.destructure.

Returns the weighted global (all batches) loss of the energies and forces.

PotentialLearning.grad_mlp — Method

grad_mlp(nn_params, x0)

nn_params: neural network parameters. x0: first layer input (energy descriptors).

Returns the analytical derivative of a feedforward neural network.

PotentialLearning.learn — Method

learn(B_train, dB_train, e_train, f_train, w_e, w_f)

B_train: energy descriptors. dB_train: force descriptors. e_train: training energies. f_train: training forces. w_e: energy weight. w_f: force weight.

PotentialLearning.linearize_forces — Method

linearize_forces(forces)

forces: vector of forces per system

Returns a vector with the components of the forces of the systems.

PotentialLearning.load_datasets — Method

load_datasets(input)

input: OrderedDict with input arguments. See get_defaults_args().

Returns training and test systems, energies, forces, and stresses.

PotentialLearning.load_extxyz — Method

load_extxyz(file; max_entries = 2000, T = Float64)

file: name of extxyz file max_entries: maximum number of systems T: default type

Returns a vector of systems or atomic configurations.

PotentialLearning.loss — Method

loss(es_pred, es, w_e, fs_pred, fs, w_f)

es_pred: preditected energies es: energies w_e: energy weight fs_pred: preditected forces fs: forces w_f: force weight

Returns the weighted loss of the energies and forces.

PotentialLearning.loss — Method

loss(x, y)

x: vector of scalars y: vector of scalars

Returns the loss of two vectors

PotentialLearning.plot_cos — Method

plot_cos(f_pred, f_true)

f_pred: vector of predicted forces f_true: vector of true forces

Returns a plot with the cosine or correlation of the forces.

PotentialLearning.plot_energy — Method

plot_energy(e_pred, e_true)

e_pred: vector of predicted energies e_true: vector of true energies

Returns an energy plot.

PotentialLearning.plot_forces — Method

plot_forces(f_pred, f_true)

f_pred: vector of predicted forces f_true: vector of true forces

Returns a force plot.

PotentialLearning.potential_energy — Method

potential_energy(A::AbstractSystem, p::NNBasisPotential)

A: atomic system or atomic configuration. p: neural network basis potential.

Returns the potential energy of a system using a neural network basis potential.

PotentialLearning.potential_energy — Method

potential_energy(b::Vector, p::NNBasisPotential)

b: vector of energy descriptors. p: neural network basis potential.

Returns the potential energy of a system using a neural network basis potential.

PotentialLearning.potential_energy — Method

potential_energy(b::Vector, ps::Vector, re)

b: energy descriptors of a system. ps: neural network parameters. See Flux.destructure. re: neural network restructure. See Flux.destructure.

Returns the potential energy of a system using a neural network basis potential.

PotentialLearning.random_subsample — Method

random_subsample(systems, energies, forces, stress, n_sys)

systems: vector with atomic configurations energies: vector with the energies of each atomic configuration forces: vector with the forces of each atomic configuration stress: vector with the stress of each atomic configuration max_sys: maximum random subset size.

Returns a random subset of the input vectors.

PotentialLearning.to_num — Method

to_num(str)

str: string with a number: integer or float

Returns an integer or float.

PotentialLearning.train! — Function

train!( train_loader_e, train_loader_f, test_loader_e, test_loader_f,
        w_e, w_f, nnbp, epochs, opt, maxiters)

train_loader_e: energy data loader for training. train_loader_f: force data loader for training. test_loader_e: energy data loader for test. test_loader_f: force data loader for test. w_e: energy weight. w_f: force weight. nnbp: neural network basis potential. epochs: no. of epochs. opt: optimizer. maxiters: maximum number of iterations in the optimizer.

Train neural network potential. Returns losses of training and test per epoch and per batch.